add network with ipaddress

+++ src/main/java/pl/lodz/p/repository/NetworkRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import pl.lodz.p.domain.entities.Network;
 
-public interface NetworkRepository extends JpaRepository<Network, Integer> {
+public interface NetworkRepository extends JpaRepository<Network, Long> {
 
 }
